Refactored lib/leave_module/services/cloud_leave_provider.dart -> lib/leave_module/services/firestore_leave_provider.dart
